refactor(alerts): address PR review comments
- Extract AlertType/AlertPayload to src/server/shared/alertTypes.ts to avoid
  type duplication between server (alert-relay, eventsub-handler) and client (useAlertStream)
- Extend overlay token on alert SSE keepalive (debounced per channel, 1h) matching chat-relay behavior
- Reset reconnectAttemptsRef before connecting when token changes in useAlertStream
- Track testAlertStatus reset timer in a ref (testAlertTimerRef) and clear on unmount
- Move VALID_ALERT_TYPES Set to module scope in server/index.ts
